refactor packfile.SeekableReader into packfile.SeekableReadRecaller and document it
