+++ storage/fs/internal/index/index.go
@@ -37,7 +37,7 @@ func NewFromIdx(r io.Reader) (Index, error) {
 func NewFromPackfile(rs io.ReadSeeker) (Index, error) {
 	index := make(Index)
 
-	r := packfile.NewSeekableReader(rs)
+	r := packfile.NewSeekableReadRecaller(rs)
 	p := packfile.NewParser(r)
 
 	count, err := p.ReadHeader()
