+++ storage/fs/store.go
@@ -125,7 +125,7 @@ func (s *ObjectStorage) Get(h core.Hash) (core.Object, error) {
 		return nil, err
 	}
 
-	r := packfile.NewSeekableReader(f)
+	r := packfile.NewSeekableReadRecaller(f)
 	r.OffsetsByHash = map[core.Hash]int64(s.index)
 	p := packfile.NewParser(r)
